frontend/requirements.txt
altair==4.2.2
astor==0.8.1
attrs==24.3.0
audeer==2.2.0
audiofile==1.5.0
audmath==1.4.1
base58==2.1.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
decorator==4.4.2
distlib==0.3.6
entrypoints==0.4
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
narwhals==1.19.0
networkx==3.2.1
nltk==3.6.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
opencv-python==4.10.0.84
packaging==24.2
pandas==1.3.0
pillow==10.4.0
proglog==0.1.10
protobuf==3.20.3
psutil==6.1.1
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
Pympler==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.0.2
scipy==1.10.1
six==1.17.0
smmap==5.0.1
soundfile==0.12.1
SpeechRecognition==3.12.0
streamlit==1.3.0
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.21.0
